+++ packages/lexical-html/src/index.ts
@@ -23,49 +23,6 @@ import {
   $sliceSelectedTextNodeContent,
 } from '@lexical/selection';
 import {$getRoot, $isElementNode, $isTextNode} from 'lexical';
-import invariant from 'shared/invariant';
-
-import {cache2Get, cache2Set} from './utils';
-
-let findParentNodeCache: null | Map<
-  Node,
-  Map<FindCachedParentDOMNodeSearchFn, null | Node>
-> = null;
-
-export type FindCachedParentDOMNode = (
-  node: Node,
-  searchFn: FindCachedParentDOMNodeSearchFn,
-) => null | Node;
-export type FindCachedParentDOMNodeSearchFn = (node: Node) => boolean;
-
-export function $findParentDOMNode(
-  node: Node,
-  searchFn: FindCachedParentDOMNodeSearchFn,
-) {
-  invariant(
-    findParentNodeCache !== null,
-    'Expected findParentNodeCache to have been initialized. This function is should be called inside importDOM. Did you call this inside $generateNodesFromDOM?',
-  );
-  let parent = node.parentNode;
-  let cached;
-  if ((cached = cache2Get(findParentNodeCache, node, searchFn)) !== undefined) {
-    return cached;
-  }
-  const visited = [node];
-  while (
-    parent !== null &&
-    (cached = cache2Get(findParentNodeCache, parent, searchFn)) === undefined &&
-    !searchFn(parent)
-  ) {
-    visited.push(parent);
-    parent = parent.parentNode;
-  }
-  const resultNode = cached === undefined ? parent : cached;
-  for (let i = 0; i < visited.length; i++) {
-    cache2Set(findParentNodeCache, visited[i], searchFn, resultNode);
-  }
-  return resultNode;
-}
 
 /**
  * How you parse your html string to get a document is left up to you. In the browser you can use the native
@@ -78,7 +35,6 @@ export function $generateNodesFromDOM(
 ): Array<LexicalNode> {
   const elements = dom.body ? dom.body.childNodes : [];
   let lexicalNodes: Array<LexicalNode> = [];
-  findParentNodeCache = new Map();
   for (let i = 0; i < elements.length; i++) {
     const element = elements[i];
     if (!IGNORE_TAGS.has(element.nodeName)) {
@@ -88,7 +44,6 @@ export function $generateNodesFromDOM(
       }
     }
   }
-  findParentNodeCache = null;
 
   return lexicalNodes;
 }
